feature: S3C-3584 restructure the unit tests by service
